% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/garfo_landings.R
\name{landings_trends}
\alias{landings_trends}
\title{Plotting landings trends}
\usage{
landings_trends(species = "all", data = "landings")
}
\arguments{
\item{species}{Mid-Atlantic managed species as listed in `mid_atlantic_species(source = "landings")`}

\item{data}{Landings outputs from `pull_garfo_landings()`}
}
\value{
List of faceted plots.
}
\description{
Function to plot landed value and volume trends for Mid-Atlantic species.
}
\examples{
# not run
}
