% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nefsc.R
\name{map_nefsc}
\alias{map_nefsc}
\title{Map NEFSC biomass data}
\usage{
map_nefsc(species = "all", data = NULL)
}
\arguments{
\item{species}{Default is "all", includes Mid-Atlantic species represented in `species.shift::species_list()`}

\item{data}{Default is "nefsc" `nefsc` must be run and named "observer" in order to run this function.}
}
\value{
Map of distribution of biomass along the Northeast US. Selecting `all` species will return a list.
}
\description{
Function to maps distributions of biomass in NEFSC Spring-Fall Bottom Trawl Survey. Color indicates `leading` and `trailing` 10% of kept catch, with `center`representing 80% of surveyed biomass. Yellow contours represent density.
}
\examples{
# map_nefsc(species = "summer flounder", data = nefsc)

}
