% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observer.R
\name{map_observer}
\alias{map_observer}
\title{Plot observer data}
\usage{
map_observer(species = "all", data = "observer")
}
\arguments{
\item{species}{Default is "all", includes Mid-Atlantic species represented in `species.shifts::species_list()`}

\item{data}{Default is "observer" `pull_observer` must be run and named "observer" in order to run this function.}
}
\value{
Map of distribution of observed catch along the Northeast US. Selecting `all` species will return a list.
}
\description{
Function to plot distributions of kept catch using federal observer data. Color indicates `leading` and `trailing` 10% of kept catch, with `center`representing 80% of kept catch. Yellow contours represent density. Distributions are cropped to management zones.
}
\examples{
# map_observer(species = "summer flounder", data = "observer")

}
