% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permits.R
\name{map_permits}
\alias{map_permits}
\title{Map permits by type along NEUS coast.}
\usage{
map_permits(species = "all", data = "permits")
}
\arguments{
\item{species}{Default is "all", includes Mid-Atlantic species represented in `species.shifts::species_list()`}

\item{data}{Default is "permits" `pull_permits` must be run and named "permits" in order to run this function.}
}
\value{
Map of distribution of observed kept catch along the Northeast US. Selecting `all` species will return a list.
}
\description{
Function to plot distributions of permits along the Northeast US coastline.
}
\examples{
# map_permits(species = "summer flounder", data = "permits")

}
