% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrip.R
\name{plot_catch_trends}
\alias{plot_catch_trends}
\title{MRIP Catch Estimate Trends}
\usage{
plot_catch_trends(species = "all", data = "NULL")
}
\arguments{
\item{species}{Mid-Atlantic managed species as listed in `species_list(source = "permits")`}

\item{data}{Landings outputs from `pull_mrip_catch()`}
}
\value{
gtable of recreational harvest trends
}
\description{
Function to calculate trends and summarise into an aesthetic table.
}
