% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nefsc.R
\name{plot_council_biomass}
\alias{plot_council_biomass}
\title{Proportions of NEFSC-surveyed biomass}
\usage{
plot_council_biomass(species = "all", data = NULL)
}
\arguments{
\item{species}{Default is "all", includes Mid-Atlantic species represented in `species.shift::species_list()`}

\item{data}{Default is "nefsc" `nefsc` must be run and named "observer" in order to run this function.}
}
\value{
Barplot of biomass proportions across Atlantic council management zones. Selecting `all` species will return a list.
}
\description{
Function to calculate and plot the proportion of surveyed biomass in each council region
}
\examples{
# plot_council_biomass(species = "summer flounder", data = nefsc)

}
