% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nefsc.R
\name{plot_nefsc_edges}
\alias{plot_nefsc_edges}
\title{Plot edges of distribution}
\usage{
plot_nefsc_edges(species = "all", data = NULL)
}
\arguments{
\item{species}{Default is "all", includes Mid-Atlantic species represented in `species.shift::species_list()`}

\item{data}{Default is "nefsc". `pull_nefsc()` must be run prior in order to run this function.}
}
\value{
Faceted plot of spring and fall percentiles of distributions. Plots both rolling mean values and smoothed values. Selecting `all` species will return a list.
}
\description{
Function to calculate and plot the 5 year rolling means of the 5ht, 50th, and 95th percentiles of biomass-weighted latitude in the NEFSC Bottom Trawl Survey. Plots spring and fall independently.
}
\examples{
# plot_nefsc_edges(species = "summer flounder", data = nefsc)
}
