% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permits.R
\name{plot_permit_edges}
\alias{plot_permit_edges}
\title{Plotting permit 5th, 50th, and 95th weighted percentiles of latitudinal distribtion.}
\usage{
plot_permit_edges(species = "all", data = "permits")
}
\arguments{
\item{species}{Mid-Atlantic managed species as listed in `species_list(source = "permits")`}

\item{data}{Landings outputs from `pull_permits()`}
}
\value{
List of faceted plots.
}
\description{
Function to plot percentiles of permits across states for Mid-Atlantic species.
}
\examples{
# plot_permit_edges(species = "summer flounder", data = permits)
}
