% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/garfo_landings.R
\name{plot_state_landings}
\alias{plot_state_landings}
\title{Plotting landings proportions by state}
\usage{
plot_state_landings(species = "all", data = "landings")
}
\arguments{
\item{species}{Mid-Atlantic managed species as listed in `species_list(source = "landings")`}

\item{data}{Landings outputs from `pull_garfo_landings()`}
}
\value{
List of faceted plots.
}
\description{
Function to plot proportions of landings across states for Mid-Atlantic species.
}
\examples{
# not run
}
