% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permits.R
\name{plot_state_permits}
\alias{plot_state_permits}
\title{Plotting permit proportions by state}
\usage{
plot_state_permits(species = "all", data = NULL)
}
\arguments{
\item{species}{Mid-Atlantic managed species as listed in `species_list(source = "permits")`}

\item{data}{Landings outputs from `pull_permits()`}
}
\value{
List of faceted plots.
}
\description{
Function to plot proportions of landings across states for Mid-Atlantic species.
}
\examples{
# plot_state_permits(species = "summer flounder", data = permits)
}
