% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrip_dirtrips.R
\name{plot_trip_estimates}
\alias{plot_trip_estimates}
\title{MRIP Directed trip estimates}
\usage{
plot_trip_estimates(species = "all", data = "NULL")
}
\arguments{
\item{species}{Mid-Atlantic managed species as listed in `species_list(source = "mrip")`}

\item{data}{Catch estimates from `pull_mrip_catch()`}
}
\value{
gtable of recreational directed trip trends
}
\description{
Function to calculate trends and summarise into an aesthetic table.
}
