% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/garfo_landings.R
\name{pull_garfo_landings}
\alias{pull_garfo_landings}
\title{Pull federal landings data}
\usage{
pull_garfo_landings(proj_path = NULL, sheet = NULL, skip = NULL)
}
\arguments{
\item{proj_path}{Local path to data file}

\item{sheet}{Derived from readxl::read_xlsx(). Indiciates which sheet of Excel file to read.}

\item{skip}{Also derived from readxl::read_xlsx(). Indicates the number of rows to skip, if any.}
}
\value{
Data frame of landings by species, principal port and state, landings, live weight, value, latitude and longitude of principal port.
}
\description{
Function to pull GARFO Dealer data and geocode principal ports from pre-existing repository.
}
\examples{
# my_path <- "/Users/clovas/Library/CloudStorage/Box-Box/CONFIDENTIAL_GARFO_MAFMC_2025/VTR_and_Dealerdata_by_port_and_species_1964-2024/KMills_VTR Data Dump & Dealer by Port_JUN 2025.xlsx"
}
