% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrip.R
\name{pull_mrip_catch}
\alias{pull_mrip_catch}
\title{Pull federal MRIP catch estimates}
\usage{
pull_mrip_catch(proj_path)
}
\arguments{
\item{proj_path}{Local path to data file}
}
\value{
Data frame of catch estimates. Please refer to [MRIP Data Dictionary](https://www.fisheries.noaa.gov/s3//2025-03/MRIP-Data-User-Handbook_March_2025_update.pdf.pdf) for more information regarding estimates.
}
\description{
Function to pull and clean recreational catch estimates. Data can be acquired from the ACCSP Data Warehouse. Please download as 'Catch_Estimates.csv'
}
\examples{
# catch <- pull_mrip_catch(proj_path = proj_path)

}
