% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observer.R
\name{pull_observer}
\alias{pull_observer}
\title{Pull federal observer data}
\usage{
pull_observer(proj_path)
}
\arguments{
\item{proj_path}{Local path to data file}
}
\value{
Data frame of observer data; contains both catch and haul information.
}
\description{
Function to pull and clean NOAA Fisheries Observer data from pre-existing confidential repository.
}
\examples{
# not run

}
