% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_list.R
\name{species_list}
\alias{species_list}
\title{Mid-Atlantic Species}
\usage{
species_list(source = "all")
}
\arguments{
\item{source}{Filter to desired data set. source = "all" returns complete list. Options include "nefsc, vtr, observer, landings, mrip, and permits"}
}
\value{
Data frame of landings by species, principal port and state, landings, live weight, value, latitude and longitude of principal port.
}
\description{
Raw and cleaned common names for species managed by the MAFMC across six federal data sets.
}
\examples{
species_list(source = "all")

}
