% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nefsc.R
\name{map_nefsc_cob}
\alias{map_nefsc_cob}
\title{Map NEFSC Bottom Trawl Center of Biomass}
\usage{
map_nefsc_cob(species = "all", data = "nefsc")
}
\arguments{
\item{species}{Default is "all", includes Mid-Atlantic species represented in `species.shift::species_list()`}

\item{data}{Default is "nefsc". `pull_nefsc()` must be run prior in order to run this function.}
}
\value{
Map of distribution of seasonal centers of biomass along the Northeast US. Selecting `all` species will return a list.
}
\description{
Function to calculate the seasonal centers of biomass of Mid-Atlantic species and map them.
}
\examples{
# map_nefsc_cob(species = "black sea bass", data = nefsc)
}
