% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtr.R
\name{map_vtr}
\alias{map_vtr}
\title{Plot vessel trip report data}
\usage{
map_vtr(species = "all", data = "vtr")
}
\arguments{
\item{species}{Default is "all", includes Mid-Atlantic species represented in `speciesshifts::mid_atlantic_species()`}

\item{data}{Default is "vtr." `pull_vtr` must be run and named "vtr" in order to run this function.}
}
\value{
Map of distribution of observed kept catch along the Northeast US. Selecting `all` species will return a list.
}
\description{
Function to plot distributions of kept catch using vessel trip report (VTR) data. Color indicates `leading` and `trailing` 10% of kept catch, with `center`representing 80% of kept catch. Yellow contours represent density. Distributions are cropped to management zones.
}
\examples{
# map_vtr(species = "summer flounder", data = "vtr")

}
