% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nefsc.R
\name{plot_nefsc_centers}
\alias{plot_nefsc_centers}
\title{Plot NEFSC Bottom Trawl Center of Biomass}
\usage{
plot_nefsc_centers(species = "all", data = "nefsc")
}
\arguments{
\item{species}{Default is "all", includes Mid-Atlantic species represented in `species.shift::species_list()`}

\item{data}{Default is "nefsc". `pull_nefsc()` must be run prior in order to run this function.}
}
\value{
Plot of fall and spring centers of latitude. Selecting `all` species will return a list.
}
\description{
Function to calculate the annual centers of biomass by season of Mid-Atlantic species and plot latitude and the distance between them.
}
\examples{
# plot_nefsc_centers(species = "black sea bass", data = nefsc)
}
