% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nefsc.R
\name{pull_nefsc}
\alias{pull_nefsc}
\title{Pull NEFSC survdat data}
\usage{
pull_nefsc(proj_path)
}
\arguments{
\item{proj_path}{Local path to data file}
}
\value{
Data frame of observer data; contains both catch and haul information.
}
\description{
Function to pull and clean NEFSC Spring-Fall Bottom Trawl Survey (survdat) data from pre-existing confidential repository.
}
\examples{
# nefsc <- pull_nefsc(proj_path = "~Data/trawl_dat.rds")
}
