% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permits.R
\name{pull_permits}
\alias{pull_permits}
\title{Pull federal permits data}
\usage{
pull_permits(proj_path)
}
\arguments{
\item{proj_path}{Local path to data file}
}
\value{
Data frame of permits; includes year, prinicpal port and state, permit type, target species, and category (commerical, for-hire).
}
\description{
Function to pull, clean and geocode GARFO commercial fishing permits from pre-existing confidential repository. Note that due to spelling error, geocoding principal ports removes 1% of permit entries, and takes approximately 25 minutes to run.
}
\examples{
# permits <- pull_permits(proj_path = proj_path)
}
