% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtr.R
\name{pull_vtr}
\alias{pull_vtr}
\title{Pull federal VTR data}
\usage{
pull_vtr(proj_path)
}
\arguments{
\item{proj_path}{Local path to data file}
}
\value{
Data frame of vessel trip reports; includes year, sub_trip_id, latitude, longitude, port name and state, species caught, weight of kept and discarded catch.
}
\description{
Function to pull and clean Vessel Trip Reports from pre-existing confidential repository.
}
\examples{
# not run

}
